+++ src/component.rs
@@ -1,19 +1,22 @@
 //! Types for working with [`Component`]s.
 
 use alloc::borrow::Cow;
-use alloc::collections::BTreeSet;
+use alloc::collections::{BTreeMap, BTreeSet};
 use core::alloc::Layout;
 use core::any::TypeId;
 use core::ops::Index;
+use core::ptr::NonNull;
 
 use ahash::RandomState;
 pub use evenio_macros::Component;
+use slab::Slab;
 
+use crate::aliased_box::AliasedBox;
 use crate::archetype::{Archetype, ArchetypeIdx};
 use crate::assert::UnwrapDebugChecked;
-use crate::drop::DropFn;
-use crate::entity::EntityLocation;
-use crate::event::{Event, EventId, EventPtr};
+use crate::drop::{drop_fn_of, DropFn};
+use crate::entity::{EntityId, EntityLocation};
+use crate::event::{Event, EventPtr};
 use crate::map::{Entry, IndexSet, TypeIdMap};
 use crate::prelude::World;
 use crate::slot_map::{Key, SlotMap};
@@ -37,14 +40,14 @@ use crate::world::UnsafeWorldCell;
 /// ```
 #[derive(Debug)]
 pub struct Components {
-    infos: SlotMap<ComponentInfo>,
-    by_type_id: TypeIdMap<ComponentId>,
+    by_index: Slab<ComponentInfoPtr>,
+    by_type_id: TypeIdMap<ComponentInfoPtr>,
 }
 
 impl Components {
     pub(crate) fn new() -> Self {
         Self {
-            infos: SlotMap::new(),
+            by_index: Slab::new(),
             by_type_id: TypeIdMap::default(),
         }
     }
@@ -53,12 +56,12 @@ impl Components {
         if let Some(type_id) = desc.type_id {
             return match self.by_type_id.entry(type_id) {
                 Entry::Vacant(v) => {
-                    let Some(k) = self.infos.insert_with(|k| ComponentInfo {
+                    let Some(k) = self.by_index.insert_with(|k| ComponentInfo {
                         name: desc.name,
                         id: ComponentId(k),
                         type_id: desc.type_id,
                         layout: desc.layout,
-                        drop: desc.drop,
+                        drop: desc.drop_fn,
                         is_immutable: desc.is_immutable,
                         insert_events: BTreeSet::new(),
                         remove_events: BTreeSet::new(),
@@ -73,12 +76,12 @@ impl Components {
             };
         }
 
-        let Some(k) = self.infos.insert_with(|k| ComponentInfo {
+        let Some(k) = self.by_index.insert_with(|k| ComponentInfo {
             name: desc.name,
             id: ComponentId(k),
             type_id: desc.type_id,
             layout: desc.layout,
-            drop: desc.drop,
+            drop: desc.drop_fn,
             is_immutable: desc.is_immutable,
             insert_events: BTreeSet::new(),
             remove_events: BTreeSet::new(),
@@ -91,7 +94,7 @@ impl Components {
     }
 
     pub(crate) fn remove(&mut self, component_id: ComponentId) -> Option<ComponentInfo> {
-        let info = self.infos.remove(component_id.0)?;
+        let info = self.by_index.remove(component_id.0)?;
 
         if let Some(type_id) = info.type_id {
             self.by_type_id.remove(&type_id);
@@ -103,17 +106,17 @@ impl Components {
     /// Gets the [`ComponentInfo`] of the given component. Returns `None` if the
     /// ID is invalid.
     pub fn get(&self, id: ComponentId) -> Option<&ComponentInfo> {
-        self.infos.get(id.0)
+        self.by_index.get(id.0)
     }
 
     /// Gets the [`ComponentInfo`] for a component using its [`ComponentIdx`].
     /// Returns `None` if the index is invalid.
     pub fn get_by_index(&self, idx: ComponentIdx) -> Option<&ComponentInfo> {
-        self.infos.get_by_index(idx.0).map(|(_, v)| v)
+        self.by_index.get_by_index(idx.0).map(|(_, v)| v)
     }
 
     pub(crate) fn get_by_index_mut(&mut self, idx: ComponentIdx) -> Option<&mut ComponentInfo> {
-        self.infos.get_by_index_mut(idx.0).map(|(_, v)| v)
+        self.by_index.get_by_index_mut(idx.0).map(|(_, v)| v)
     }
 
     /// Gets the [`ComponentInfo`] for a component using its [`TypeId`]. Returns
@@ -130,7 +133,7 @@ impl Components {
 
     /// Returns an iterator over all component infos.
     pub fn iter(&self) -> impl Iterator<Item = &ComponentInfo> {
-        self.infos.iter().map(|(_, v)| v)
+        self.by_index.iter().map(|(_, v)| v)
     }
 }
 
@@ -194,70 +197,112 @@ unsafe impl SystemParam for &'_ Components {
     fn remove_archetype(_state: &mut Self::State, _arch: &Archetype) {}
 }
 
-/// Metadata for a component.
-#[derive(Debug)]
-pub struct ComponentInfo {
-    name: Cow<'static, str>,
-    id: ComponentId,
-    type_id: Option<TypeId>,
-    layout: Layout,
-    drop: DropFn,
-    is_immutable: bool,
-    pub(crate) insert_events: BTreeSet<EventId>,
-    pub(crate) remove_events: BTreeSet<EventId>,
-    /// The set of archetypes that have this component as one of its columns.
-    pub(crate) member_of: IndexSet<ArchetypeIdx>,
-}
+#[derive(Component, Debug)]
+#[component(immutable)] // Required for soundness.
+#[repr(transparent)]
+pub struct ComponentInfo(AliasedBox<ComponentInfoInner>);
 
 impl ComponentInfo {
-    /// Gets the name of the component.
-    ///
-    /// This name is intended for debugging purposes and should not be relied
-    /// upon for correctness.
-    pub fn name(&self) -> &str {
-        &self.name
+    pub fn new<C: Component>() -> Self {
+        unsafe {
+            Self::from_descriptor(ComponentDescriptor {
+                type_id: Some(TypeId::of::<C>()),
+                layout: Layout::new::<C>(),
+                drop_fn: drop_fn_of::<C>(),
+                is_immutable: false,
+            })
+        }
     }
 
-    /// Gets the ID of the component.
-    pub fn id(&self) -> ComponentId {
-        self.id
+    pub unsafe fn from_descriptor(
+        ComponentDescriptor {
+            type_id,
+            layout,
+            drop_fn: drop,
+            is_immutable,
+        }: ComponentDescriptor,
+    ) -> Self {
+        Self(AliasedBox::new(ComponentInfoInner {
+            entity_id: EntityId::NULL,
+            type_id,
+            layout,
+            drop_fn: drop,
+            is_immutable,
+            insert_events: BTreeSet::new(),
+            remove_events: BTreeSet::new(),
+        }))
     }
 
-    /// Gets the [`TypeId`] of the component, or `None` if it was not assigned a
-    /// type ID.
+    pub fn entity_id(&self) -> EntityId {
+        self.0.entity_id
+    }
+
     pub fn type_id(&self) -> Option<TypeId> {
-        self.type_id
+        self.0.type_id
     }
 
-    /// Gets the [`Layout`] of the component.
     pub fn layout(&self) -> Layout {
-        self.layout
+        self.0.layout
     }
 
-    /// Gets the [`DropFn`] of the component.
-    pub fn drop(&self) -> DropFn {
-        self.drop
+    pub fn drop_fn(&self) -> DropFn {
+        self.0.drop_fn
     }
 
-    /// Gets the [immutability] of the component.
-    ///
-    /// [immutability]: Component::IS_IMMUTABLE
     pub fn is_immutable(&self) -> bool {
-        self.is_immutable
+        self.0.is_immutable
+    }
+}
+
+#[derive(Debug)]
+struct ComponentInfoInner {
+    entity_id: EntityId,
+    type_id: Option<TypeId>,
+    layout: Layout,
+    drop_fn: DropFn,
+    is_immutable: bool,
+    insert_events: BTreeSet<EntityId>,
+    remove_events: BTreeSet<EntityId>,
+}
+
+/// Data needed to create a new component.
+#[derive(Clone, Debug)]
+pub struct ComponentDescriptor {
+    /// The [`TypeId`] of this component, if any.
+    pub type_id: Option<TypeId>,
+    /// The [`Layout`] of the component.
+    pub layout: Layout,
+    /// The [`DropFn`] of the component. This is passed a pointer to the
+    /// component in order to drop it.
+    pub drop_fn: DropFn,
+    /// If this component is [immutable](Component::IS_IMMUTABLE).
+    pub is_immutable: bool,
+}
+
+/// Pointer to component info.
+#[derive(Clone, Copy, Debug)]
+#[repr(transparent)]
+pub(crate) struct ComponentInfoPtr(NonNull<ComponentInfoInner>);
+
+impl ComponentInfoPtr {
+    /// # Safety
+    ///
+    /// - Pointer must be valid.
+    /// - Aliasing rules must be followed.
+    pub(crate) unsafe fn as_info(&self) -> &ComponentInfo {
+        // SAFETY: Both `ComponentInfo` and `ComponentInfoPtr` have non-null pointer
+        // layout.
+        &*(self as *const Self as *const ComponentInfo)
     }
 
-    /// Gets the set of [`Insert`] events for this component.
+    /// # Safety
     ///
-    /// [`Insert`]: crate::event::Insert
-    pub fn insert_events(&self) -> &BTreeSet<EventId> {
-        &self.insert_events
-    }
-
-    /// Gets the set of [`Remove`] components for this component.
-    ///
-    /// [`Remove`]: crate::event::Remove
-    pub fn remove_events(&self) -> &BTreeSet<EventId> {
-        &self.remove_events
+    /// - Pointer must be valid.
+    /// - Aliasing rules must be followed.
+    pub(crate) unsafe fn as_info_mut(&mut self) -> &mut ComponentInfo {
+        // SAFETY: Both `ComponentInfo` and `ComponentInfoPtr` have non-null pointer
+        // layout.
+        &mut *(self as *mut Self as *mut ComponentInfo)
     }
 }
 
@@ -320,65 +365,6 @@ pub trait Component: Send + Sync + 'static {
     const IS_IMMUTABLE: bool = false;
 }
 
-/// Data needed to create a new component.
-#[derive(Clone, Debug)]
-pub struct ComponentDescriptor {
-    /// The name of this component.
-    ///
-    /// This name is intended for debugging purposes and should not be relied
-    /// upon for correctness.
-    pub name: Cow<'static, str>,
-    /// The [`TypeId`] of this component, if any.
-    pub type_id: Option<TypeId>,
-    /// The [`Layout`] of the component.
-    pub layout: Layout,
-    /// The [`DropFn`] of the component. This is passed a pointer to the
-    /// component in order to drop it.
-    pub drop: DropFn,
-    /// If this component is [immutable](Component::IS_IMMUTABLE).
-    pub is_immutable: bool,
-}
-
-/// Lightweight identifier for a component type.
-///
-/// component identifiers are implemented using an [index] and a generation
-/// count. The generation count ensures that IDs from removed components are
-/// not reused by new components.
-///
-/// A component identifier is only meaningful in the [`World`] it was created
-/// from. Attempting to use a component ID in a different world will have
-/// unexpected results.
-///
-/// [index]: ComponentIdx
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Debug)]
-pub struct ComponentId(Key);
-
-impl ComponentId {
-    /// The component ID which never identifies a live component. This is the
-    /// default value for `ComponentId`.
-    pub const NULL: Self = Self(Key::NULL);
-
-    /// Creates a new component ID from an index and generation count. Returns
-    /// `None` if a valid ID is not formed.
-    pub const fn new(index: u32, generation: u32) -> Option<Self> {
-        match Key::new(index, generation) {
-            Some(k) => Some(Self(k)),
-            None => None,
-        }
-    }
-
-    /// Returns the index of this ID.
-    pub const fn index(self) -> ComponentIdx {
-        ComponentIdx(self.0.index())
-    }
-
-    /// Returns the generation count of this ID.
-    pub const fn generation(self) -> u32 {
-        self.0.generation().get()
-    }
-}
-
-/// A [`ComponentId`] with the generation count stripped out.
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Debug)]
 pub struct ComponentIdx(pub u32);
 
@@ -393,16 +379,6 @@ unsafe impl SparseIndex for ComponentIdx {
         Self(u32::from_index(idx))
     }
 }
-
-/// An event sent immediately after a new component is added to the world.
-/// Contains the ID of the added component.
-#[derive(Event, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub struct AddComponent(pub ComponentId);
-
-/// An event sent immediately before a component is removed from the world.
-/// Contains the ID of the component to be removed.
-#[derive(Event, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub struct RemoveComponent(pub ComponentId);
 
 #[cfg(test)]
 mod tests {
@@ -446,7 +422,7 @@ mod tests {
             b.push("hello".into());
         });
         world.send(E);
-        assert_eq!(world.get_component::<B>(e2), Some(&B(vec!["hello".into()])));
+        assert_eq!(world.get::<B>(e2), Some(&B(vec!["hello".into()])));
 
         assert!(world.remove_component(c2).is_some());
         assert!(!world.systems().contains(s2));