+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Removed `Column::component_index`. To obtain the component index of a column, use `Archetype::component_indices`.
 - Fixed bug related to system ordering. Systems with the same priority always run in the order they were added to the world.
+- Renamed `World::get_component(_mut)` to `World::get(_mut)`.
 
 ## 0.3.0 - 2024-02-13